Move parseStackTrace to ReactFlightServerConfig
All servers currently use the V8 format. Including Bun, which overrides
the JSC format to V8 format it seems.

However, we might need to configure this for the FB build. Conceptually it's
also just something that is environment specific.

I now apply filtering on the result which needs to update some hacks.
