+++ packages/react-server/src/forks/ReactFlightServerConfig.custom.js
@@ -14,6 +14,8 @@ export * from '../ReactFlightServerConfigBundlerCustom';
 
 export * from '../ReactFlightServerConfigDebugNoop';
 
+export * from '../ReactFlightStackConfigV8';
+
 export type Hints = any;
 export type HintCode = any;
 // eslint-disable-next-line no-unused-vars
