+++ packages/react-server/src/forks/ReactFlightServerConfig.dom-edge-turbopack.js
@@ -35,4 +35,7 @@ export const createAsyncHook: HookCallbacks => AsyncHook =
       };
 export const executionAsyncId: () => number =
   typeof async_hooks === 'object' ? async_hooks.executionAsyncId : (null: any);
+
 export * from '../ReactFlightServerConfigDebugNode';
+
+export * from '../ReactFlightStackConfigV8';
