+++ packages/react-server/src/forks/ReactFlightServerConfig.dom-edge.js
@@ -36,4 +36,7 @@ export const createAsyncHook: HookCallbacks => AsyncHook =
       };
 export const executionAsyncId: () => number =
   typeof async_hooks === 'object' ? async_hooks.executionAsyncId : (null: any);
+
 export * from '../ReactFlightServerConfigDebugNode';
+
+export * from '../ReactFlightStackConfigV8';
