+++ packages/react-server/src/forks/ReactFlightServerConfig.markup.js
@@ -28,6 +28,8 @@ export const componentStorage: AsyncLocalStorage<ReactComponentInfo | void> =
 
 export * from '../ReactFlightServerConfigDebugNoop';
 
+export * from '../ReactFlightStackConfigV8';
+
 export type ClientManifest = null;
 export opaque type ClientReference<T> = null; // eslint-disable-line no-unused-vars
 export opaque type ServerReference<T> = null; // eslint-disable-line no-unused-vars
