+++ scripts/jest/setupTests.js
@@ -292,13 +292,13 @@ function lazyRequireFunctionExports(moduleName) {
         // If this export is a function, return a wrapper function that lazily
         // requires the implementation from the current module cache.
         if (typeof originalModule[prop] === 'function') {
-          const wrapper = function () {
-            return jest.requireActual(moduleName)[prop].apply(this, arguments);
-          };
-          // We use this to trick the filtering of Flight to exclude this frame.
-          Object.defineProperty(wrapper, 'name', {
-            value: '(<anonymous>)',
-          });
+          // eslint-disable-next-line no-eval
+          const wrapper = eval(`
+            (function () {
+              return jest.requireActual(moduleName)[prop].apply(this, arguments);
+            })
+            // We use this to trick the filtering of Flight to exclude this frame.
+            //# sourceURL=<anonymous>`);
           return wrapper;
         } else {
           return originalModule[prop];
